+++ packages/react-interactions/accessibility/src/__tests__/FocusTable-test.internal.js
@@ -7,7 +7,7 @@
  * @flow
  */
 
-import {createEventTarget} from 'react-interactions/events/src/dom/testing-library';
+import {createEventTarget} from 'react-interactions/events/src/dom/event-testing-library';
 import {emulateBrowserTab} from '../shared/emulateBrowserTab';
 
 let React;
