+++ packages/react-interactions/events/src/dom/__tests__/PressLegacy-test.internal.js
@@ -13,8 +13,9 @@ import {
   buttonType,
   buttonsType,
   createEventTarget,
+  resetActivePointers,
   setPointerEvent,
-} from '../testing-library';
+} from '../event-testing-library';
 
 let React;
 let ReactFeatureFlags;
@@ -59,6 +60,7 @@ describe.each(environmentTable)('Press responder', hasPointerEvents => {
     ReactDOM.render(null, container);
     document.body.removeChild(container);
     container = null;
+    resetActivePointers();
   });
 
   describe('disabled', () => {
